requirements.txt
streamlit
pandas
numpy
torch
torchvision
matplotlib
seaborn
pillow
pytorch-lightning